Crate servers

Modules

configurator

error

export_metrics

grpc

heartbeat_options

http

influxdb

interceptor

line_writer

metrics 🔒

metrics_handler

mysql

opentsdb

otlp

postgres

prom_row_builder 🔒

prom_store

prometheus protocol supportings handles prometheus remote_write, remote_read logic

prometheus_handler

prom supply the prometheus HTTP API Server compliance

proto

query_handler

All query handler traits for various request protocols, like SQL or GRPC. Instance that wishes to support certain request protocol, just implement the corresponding trait, the Server will handle codec for you.

repeated_field 🔒

row_writer 🔒

server

shutdown 🔒

Copied from tokio’s mini-redis example.

tls

Macros

add_service

Add a gRPC service (service) to a builder(RoutesBuilder). This macro will automatically add some gRPC properties to the service.

define_into_tonic_status

Structs

SqlPlan

Cached SQL and logical plan for database interfaces

Enums

Mode