servers::grpc

Module context_auth

Functions

auth

Authenticate the user based on the header and query context.

check_auth

Helper function to extract the header from the metadata and authenticate the user.

create_query_context_from_grpc_metadata

Create a query context from the grpc metadata.

extract_header

Helper function to extract a header from the metadata map. Can be multiple keys, and the first one found will be returned.