servers::http
Module event

Structs

JSON_CONTENT_TYPE

LogIngestRequest 🔒

LogIngestRequest is the internal request for log ingestion. The raw log input can be transformed into multiple LogIngestRequests. Multiple LogIngestRequests will be ingested into the same database with the same pipeline.

LogIngesterQueryParams

LogIngesterQueryParams is used for query params of log ingester API.

LogState

axum state struct to hold log handler and validator

PB_CONTENT_TYPE

PipelineContent

PipelineDryrunParams

Dryrun pipeline with given data pipeline_name and pipeline_version to specify pipeline stored in db pipeline to specify pipeline raw content data to specify data data maght be list of string or list of object

TEXT_CONTENT_TYPE

TEXT_UTF8_CONTENT_TYPE

Constants

GREPTIME_INTERNAL_IDENTITY_PIPELINE_NAME

GREPTIME_INTERNAL_PIPELINE_NAME_PREFIX 🔒

Traits

LogValidator

Functions

add_pipeline

add_step_info_for_pipeline_dryrun_error 🔒

check_data_valid 🔒

Check if the data length less than 10

check_pipeline_dryrun_params_valid 🔒

Check if the payload is valid json Check if the payload contains pipeline or pipeline_name and data Return Some if valid, None if invalid

check_pipeline_name_exists 🔒

Check if the pipeline_name exists

delete_pipeline

dryrun_pipeline_inner 🔒

Dryrun pipeline with given data

extract_pipeline_value_by_content_type 🔒

ingest_logs_inner 🔒

log_ingester

pipeline_dryrun

transform_ndjson_array_factory 🔒

Transform NDJSON array into a single array always return an array

Type Aliases

LogValidatorRef