servers::http
Module handler

Structs

HealthQuery

HealthResponse

PromqlQuery

SqlQuery

StatusResponse

Functions

config

Handler to expose configuration information info about runtime, build, etc.

from_output

Create a response from query result

health

Handler to export healthy check

metrics

Handler to export metrics

promql

Handler to execute promql

sql

Handler to execute sql

sql_parse

Handler to parse sql

status

Handler to expose information info about runtime, build, etc.

validate_schema 🔒