servers::http
Module jaeger

Structs

FIND_TRACES_COLS

FIND_TRACES_SCHEMA 🔒

JaegerAPIError

JaegerAPIError is the error of Jaeger HTTP API.

JaegerAPIResponse

JaegerAPIResponse is the response of Jaeger HTTP API. The original version is structuredResponse which is defined in https://github.com/jaegertracing/jaeger/blob/main/cmd/query/app/http_handler.go.

JaegerQueryParams

JaegerQueryParams is the query parameters of Jaeger HTTP API.

KeyValue

KeyValue is a key-value pair with typed value.

Log

Log is a log emitted in a span.

Operation

Operation is an operation in a service.

Process

Process is the process emitting a set of spans.

QueryTraceParams

Reference

Reference is a reference from one span to another.

Span

Span is a single operation within a trace.

Trace

Trace is a collection of spans.

Enums

JaegerData

JaegerData is the query result of Jaeger HTTP API.

Value

Value is the value of a key-value pair in Jaeger Span attributes.

ValueType

ValueType is the type of a value stored in KeyValue struct.

Constants

REF_TYPE_CHILD_OF 🔒

SPAN_KIND_TIME_FMTS 🔒

Functions

check_schema 🔒

convert_string_to_boolean 🔒

convert_string_to_number 🔒

covert_to_records 🔒

error_response 🔒

handle_find_traces

Handle the GET /api/traces request.

handle_get_operations

Handle the GET /api/operations request.

handle_get_operations_by_service

Handle the GET /api/services/{service_name}/operations request.

handle_get_services

Handle the GET /api/services request.

handle_get_trace

Handle the GET /api/traces/{trace_id} request.

handle_query_error 🔒

normalize_span_kind 🔒

normalize_status_code 🔒

object_to_tags 🔒

operations_from_records 🔒

services_from_records 🔒

traces_from_records 🔒