servers::http

Module prometheus

prom supply the prometheus HTTP API Server compliance

Macros

try_call_return_response 🔒

Helper macro which try to evaluate the expression and return its results. If the evaluation fails, return a PrometheusJsonResponse early.

Structs

BuildInfoQuery

FormatQuery

InstantQuery

LabelValueQuery

LabelsQuery

Matches 🔒

ParseQuery

PromData

PromSeriesMatrix

For [ValueType::Matrix] result type

PromSeriesVector

For [ValueType::Vector] result type

PrometheusJsonResponse

RangeQuery

SeriesQuery

Enums

PromQueryResult

Variants corresponding to [ValueType]

PrometheusResponse

Functions

build_info_query

do_instant_query 🔒

Executes a single instant query and returns response

do_range_query 🔒

Executes a single range query and returns response

find_metric_name_and_matchers 🔒

find_metric_name_not_equal_matchers 🔒

Try to find the __name__ matchers which op is not MatchOp::Equal.

format_query

get_all_column_names 🔒

Get all tag column name of the given schema

get_catalog_schema 🔒

Tries to get catalog and schema from an optional db param. And retrieves them from QueryContext if they don’t present.

instant_query

label_values_query

labels_query

parse_query

promql_expr_to_metric_name 🔒

range_query

record_batches_to_labels_name 🔒

Retrieve labels name from record batches

record_batches_to_series 🔒

retrieve_field_names 🔒

retrieve_labels_name_from_query_result 🔒

Retrieve labels name from query result

retrieve_metric_name_and_result_type 🔒

retrieve_metric_name_from_promql 🔒

Try to parse and extract the name of referenced metric from the promql query.

retrieve_series_from_query_result 🔒

series_query

take_metric_name 🔒

Take metric name from the [VectorSelector]. It takes the name in the selector or removes the name matcher.

try_update_catalog_schema 🔒

Update catalog and schema in QueryContext if necessary.

update_metric_name_matcher 🔒

Update the __name__ matchers in expression into special value Returns the updated expression.