servers::http::prometheus

Macro handle_schema_err

macro_rules! handle_schema_err {
    ($result:expr) => { ... };
}

Handles schema errors, transforming a Result into an Option.

If the input is Ok(v), returns Some(v)

If the input is Err(err) and the error status code is TableNotFound or TableColumnNotFound, returns None (ignoring these specific errors)

If the input is Err(err) with any other error code, directly returns a PrometheusJsonResponse::error.