Module servers::interceptor

Traits

GrpcQueryInterceptor

GrpcQueryInterceptor can track life cycle of a grpc request and customize or abort its execution at given point.

LineProtocolInterceptor

LineProtocolInterceptor can track life cycle of a line protocol request and customize or abort its execution at given point.

OpenTelemetryProtocolInterceptor

OpenTelemetryProtocolInterceptor can track life cycle of an open telemetry protocol request and customize or abort its execution at given point.

PromQueryInterceptor

PromQueryInterceptor can track life cycle of a prometheus request and customize or abort its execution at given point.

PromStoreProtocolInterceptor

PromStoreProtocolInterceptor can track life cycle of a prom store request and customize or abort its execution at given point.

ScriptInterceptor

ScriptInterceptor can track life cycle of a script request and customize or abort its execution at given point.

SqlQueryInterceptor

SqlQueryInterceptor can track life cycle of a sql query and customize or abort its execution at given point.

Type Aliases

GrpcQueryInterceptorRef

LineProtocolInterceptorRef

OpenTelemetryProtocolInterceptorRef

PromQueryInterceptorRef

PromStoreProtocolInterceptorRef

ScriptInterceptorRef

SqlQueryInterceptorRef