servers::interceptor

Trait GrpcQueryInterceptor

pub trait GrpcQueryInterceptor {
    type Error: ErrorExt;

    // Provided methods
    fn pre_execute(
        &self,
        _request: &Request,
        _query_ctx: QueryContextRef,
    ) -> Result<(), Self::Error> { ... }
    fn pre_bulk_insert(
        &self,
        _table: TableRef,
        _query_ctx: QueryContextRef,
    ) -> Result<(), Self::Error> { ... }
    fn post_execute(
        &self,
        output: Output,
        _query_ctx: QueryContextRef,
    ) -> Result<Output, Self::Error> { ... }
}

GrpcQueryInterceptor can track life cycle of a grpc request and customize or abort its execution at given point.

Required Associated Types

type Error: ErrorExt

Provided Methods

fn pre_execute( &self, _request: &Request, _query_ctx: QueryContextRef, ) -> Result<(), Self::Error>

Called before request is actually executed.

fn pre_bulk_insert( &self, _table: TableRef, _query_ctx: QueryContextRef, ) -> Result<(), Self::Error>

Called before bulk insert is executed.

fn post_execute( &self, output: Output, _query_ctx: QueryContextRef, ) -> Result<Output, Self::Error>

Called after execution finished. The implementation can modify the output if needed.