servers::interceptor

Trait OpenTelemetryProtocolInterceptor

pub trait OpenTelemetryProtocolInterceptor {
    type Error: ErrorExt;

    // Provided method
    fn pre_execute(
        &self,
        _query_ctx: QueryContextRef,
    ) -> Result<(), Self::Error> { ... }
}

OpenTelemetryProtocolInterceptor can track life cycle of an open telemetry protocol request and customize or abort its execution at given point.

Required Associated Types

type Error: ErrorExt

Provided Methods

fn pre_execute(&self, _query_ctx: QueryContextRef) -> Result<(), Self::Error>