Trait servers::interceptor::PromQueryInterceptor

pub trait PromQueryInterceptor {
    type Error: ErrorExt;

    // Provided methods
    fn pre_execute(
        &self,
        _query: &PromQuery,
        _query_ctx: QueryContextRef
    ) -> Result<(), Self::Error> { ... }
    fn post_execute(
        &self,
        output: Output,
        _query_ctx: QueryContextRef
    ) -> Result<Output, Self::Error> { ... }
}

PromQueryInterceptor can track life cycle of a prometheus request and customize or abort its execution at given point.

Required Associated Types

type Error: ErrorExt

Provided Methods

fn pre_execute( &self, _query: &PromQuery, _query_ctx: QueryContextRef ) -> Result<(), Self::Error>

Called before request is actually executed.

fn post_execute( &self, output: Output, _query_ctx: QueryContextRef ) -> Result<Output, Self::Error>

Called after execution finished. The implementation can modify the output if needed.