Module servers::prom_store

prometheus protocol supportings handles prometheus remote_write, remote_read logic

Structs

Metrics

Metrics for push gateway protocol

TimeSeriesId 🔒

Constants

FIELD_NAME_LABEL

The same as FIELD_COLUMN_MATCHER in promql crate

METRIC_NAME_LABEL

METRIC_NAME_LABEL_BYTES

Functions

collect_timeseries_ids 🔒

Collect each row’s timeseries id This processing is ugly, hope https://github.com/GreptimeTeam/greptimedb/issues/336 making some progress in future.

lit_timestamp_millisecond 🔒

mock_timeseries

Mock timeseries for test, it is both used in servers and frontend crate So we present it here

new_label 🔒

query_to_plan

Create a DataFrame from a remote Query

recordbatch_to_timeseries 🔒

recordbatches_to_timeseries

snappy_compress

snappy_decompress

table_name

Get table name from remote query

to_grpc_row_insert_requests

zstd_decompress