servers

Module proto

Structs

PromLabel

PromSeriesProcessor

A hook to be injected into the PromWriteRequest decoding process. It was originally designed with two usage:

PromTimeSeries

PromWriteRequest

Functions

copy_to_bytes 🔒

decode_string 🔒

Decodes bytes into String values according provided validation mode.

merge_bytes 🔒

Reads a variable-length encoded bytes field from buf and assign it to value.

split_to 🔒

Similar to Bytes::split_to, but directly operates on underlying memory region.