servers::query_handler

Trait JaegerQueryHandler

pub trait JaegerQueryHandler {
    // Required methods
    fn get_services<'life0, 'async_trait>(
        &'life0 self,
        ctx: QueryContextRef,
    ) -> Pin<Box<dyn Future<Output = Result<Output>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_operations<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        ctx: QueryContextRef,
        service_name: &'life1 str,
        span_kind: Option<&'life2 str>,
        start_time: Option<i64>,
        end_time: Option<i64>,
    ) -> Pin<Box<dyn Future<Output = Result<Output>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn get_trace<'life0, 'life1, 'async_trait>(
        &'life0 self,
        ctx: QueryContextRef,
        trace_id: &'life1 str,
        start_time: Option<i64>,
        end_time: Option<i64>,
    ) -> Pin<Box<dyn Future<Output = Result<Output>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn find_traces<'life0, 'async_trait>(
        &'life0 self,
        ctx: QueryContextRef,
        query_params: QueryTraceParams,
    ) -> Pin<Box<dyn Future<Output = Result<Output>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Handle Jaeger query requests.

Required Methods

fn get_services<'life0, 'async_trait>( &'life0 self, ctx: QueryContextRef, ) -> Pin<Box<dyn Future<Output = Result<Output>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get trace services. It’s used for /api/services API.

fn get_operations<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, ctx: QueryContextRef, service_name: &'life1 str, span_kind: Option<&'life2 str>, start_time: Option<i64>, end_time: Option<i64>, ) -> Pin<Box<dyn Future<Output = Result<Output>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Get Jaeger operations. It’s used for /api/operations and /api/services/{service_name}/operations API.

fn get_trace<'life0, 'life1, 'async_trait>( &'life0 self, ctx: QueryContextRef, trace_id: &'life1 str, start_time: Option<i64>, end_time: Option<i64>, ) -> Pin<Box<dyn Future<Output = Result<Output>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Retrieves a trace by its unique identifier.

This method is used to handle requests to the /api/traces/{trace_id} endpoint. It accepts optional start_time and end_time parameters in nanoseconds to filter the trace data within a specific time range.

fn find_traces<'life0, 'async_trait>( &'life0 self, ctx: QueryContextRef, query_params: QueryTraceParams, ) -> Pin<Box<dyn Future<Output = Result<Output>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Find traces by query params. It’s used for /api/traces API.