servers::query_handler
Trait PipelineHandler

pub trait PipelineHandler {
    // Required methods
    fn insert<'life0, 'async_trait>(
        &'life0 self,
        input: RowInsertRequests,
        ctx: QueryContextRef,
    ) -> Pin<Box<dyn Future<Output = Result<Output>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_pipeline<'life0, 'life1, 'async_trait>(
        &'life0 self,
        name: &'life1 str,
        version: PipelineVersion,
        query_ctx: QueryContextRef,
    ) -> Pin<Box<dyn Future<Output = Result<Arc<Pipeline<GreptimeTransformer>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn insert_pipeline<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        name: &'life1 str,
        content_type: &'life2 str,
        pipeline: &'life3 str,
        query_ctx: QueryContextRef,
    ) -> Pin<Box<dyn Future<Output = Result<PipelineInfo>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait;
    fn delete_pipeline<'life0, 'life1, 'async_trait>(
        &'life0 self,
        name: &'life1 str,
        version: PipelineVersion,
        query_ctx: QueryContextRef,
    ) -> Pin<Box<dyn Future<Output = Result<Option<()>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_table<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        table: &'life1 str,
        query_ctx: &'life2 QueryContext,
    ) -> Pin<Box<dyn Future<Output = Result<Option<Arc<Table>>, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn build_pipeline(
        &self,
        pipeline: &str,
    ) -> Result<Pipeline<GreptimeTransformer>>;
}

PipelineHandler is responsible for handling pipeline related requests.

The “Pipeline” is a series of transformations that can be applied to unstructured data like logs. This handler is responsible to manage pipelines and accept data for processing.

The pipeline is stored in the database and can be retrieved by its name.

Required Methods

fn insert<'life0, 'async_trait>( &'life0 self, input: RowInsertRequests, ctx: QueryContextRef, ) -> Pin<Box<dyn Future<Output = Result<Output>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn get_pipeline<'life0, 'life1, 'async_trait>( &'life0 self, name: &'life1 str, version: PipelineVersion, query_ctx: QueryContextRef, ) -> Pin<Box<dyn Future<Output = Result<Arc<Pipeline<GreptimeTransformer>>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn insert_pipeline<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 self, name: &'life1 str, content_type: &'life2 str, pipeline: &'life3 str, query_ctx: QueryContextRef, ) -> Pin<Box<dyn Future<Output = Result<PipelineInfo>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

fn delete_pipeline<'life0, 'life1, 'async_trait>( &'life0 self, name: &'life1 str, version: PipelineVersion, query_ctx: QueryContextRef, ) -> Pin<Box<dyn Future<Output = Result<Option<()>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn get_table<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, table: &'life1 str, query_ctx: &'life2 QueryContext, ) -> Pin<Box<dyn Future<Output = Result<Option<Arc<Table>>, Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

fn build_pipeline( &self, pipeline: &str, ) -> Result<Pipeline<GreptimeTransformer>>