Crate servers

Modules

addrs

configurator

elasticsearch 🔒

error

grpc

hint_headers 🔒

http

influxdb

interceptor

metrics

metrics_handler

mysql

opentsdb

otel_arrow

otlp

pending_rows_batcher

pipeline 🔒

postgres

prom_remote_write

Prometheus remote write support.

prom_row_builder 🔒

Prometheus row-level helpers for converting proto Rows into Arrow RecordBatches and aligning / normalizing their schemas against existing table schemas in the catalog.

prom_store

prometheus protocol supportings handles prometheus remote_write, remote_read logic

prometheus

prometheus_handler

prom supply the prometheus HTTP API Server compliance

query_handler

All query handler traits for various request protocols, like SQL or GRPC.

repeated_field

request_memory_limiter

Unified memory limiter for all server request protocols.

request_memory_metrics

Unified metrics adapter for all server protocols.

row_writer 🔒

server

tls

Macros

add_service

Add a gRPC service (service) to a builder([RoutesBuilder]). This macro will automatically add some gRPC properties to the service.

Structs

SqlPlan

Cached SQL and logical plan for database interfaces

Functions

install_ring_crypto_provider

Install the ring crypto provider for rustls process-wide. see: