Crate servers

Modules

auth

configurator

error

grpc

heartbeat_options

http

influxdb

interceptor

line_writer

metrics 🔒

metrics_handler

mysql

opentsdb

otlp

postgres

prom_store

prometheus protocol supportings handles prometheus remote_write, remote_read logic

prometheus

prom supply the prometheus HTTP API Server compliance

query_handler

All query handler traits for various request protocols, like SQL or GRPC. Instance that wishes to support certain request protocol, just implement the corresponding trait, the Server will handle codec for you.

server

shutdown 🔒

Copied from tokio’s mini-redis example.

tls

Structs

SqlPlan

Cached SQL and logical plan for database interfaces

Enums

Mode