servers::batcher

Module logical_table

Modules

batch 🔒

batch_convert 🔒

flow_notifier 🔒

pending_worker 🔒

region_write 🔒

tables 🔒

Structs

BatchKey 🔒

LogicalTablePendingRowsBatcher

Prometheus remote write pending rows batcher.

PhysicalTableMetadata

RecordBatchWithTsIdx

An aligned logical record batch and its timestamp column index.

TableBatch

Constants

PHYSICAL_TABLE_KEY 🔒

WORKER_IDLE_TIMEOUT_MULTIPLIER 🔒

Traits

PendingRowsSchemaAlterer

PhysicalFlushCatalogProvider

PhysicalFlushNodeRequester

PhysicalFlushPartitionProvider

Functions

batch_key_from_ctx 🔒

flush_batch_physical

Flushes a batch of logical table rows by transforming them into the physical table format and writing them to the appropriate datanode regions.

Type Aliases

PendingRowsSchemaAltererRef