servers::http
Module extractor

Structs

LogTableName

Axum extractor for optional target log table name from HTTP header using GREPTIME_LOG_TABLE_NAME_HEADER_NAME as key.

PipelineInfo

Axum extractor for optional Pipeline name and version from HTTP headers.

SelectInfoWrapper

Axum extractor for select keys from HTTP header, to extract and uplift key-values from OTLP attributes. See [SelectInfo] for more details.

TraceTableName

Axum extractor for optional target trace table name from HTTP header using GREPTIME_TRACE_TABLE_NAME_HEADER_NAME as key.

Functions

string_value_from_header 🔒