Module servers::http::prometheus

prom supply the prometheus HTTP API Server compliance

Structs

BuildInfoQuery

FormatQuery

InstantQuery

LabelValueQuery

LabelsQuery

Matches 🔒

PromData

PromSeriesMatrix

For [ValueType::Matrix] result type

PromSeriesVector

For [ValueType::Vector] result type

PrometheusJsonResponse

RangeQuery

SeriesQuery

Enums

PromQueryResult

Variants corresponding to [ValueType]

PrometheusResponse

Functions

build_info_query

format_query

get_all_column_names 🔒

Get all tag column name of the given schema

get_catalog_schema 🔒

Tries to get catalog and schema from an optional db param. And retrieves them from [QueryContext] if they don’t present.

instant_query

label_values_query

labels_query

promql_expr_to_metric_name 🔒

range_query

record_batches_to_labels_name 🔒

Retrieve labels name from record batches

record_batches_to_series 🔒

retrieve_field_names 🔒

retrieve_label_values 🔒

retrieve_label_values_from_record_batch 🔒

retrieve_labels_name_from_query_result 🔒

Retrieve labels name from query result

retrieve_metric_name_and_result_type 🔒

retrieve_metric_name_from_promql 🔒

Try to parse and extract the name of referenced metric from the promql query.

retrieve_series_from_query_result 🔒

series_query

try_update_catalog_schema 🔒

Update catalog and schema in [QueryContext] if necessary.