servers::interceptor

Trait PromStoreProtocolInterceptor

pub trait PromStoreProtocolInterceptor {
    type Error: ErrorExt;

    // Provided methods
    fn pre_write(
        &self,
        _write_req: &RowInsertRequests,
        _ctx: QueryContextRef,
    ) -> Result<(), Self::Error> { ... }
    fn pre_read(
        &self,
        _read_req: &ReadRequest,
        _ctx: QueryContextRef,
    ) -> Result<(), Self::Error> { ... }
}

PromStoreProtocolInterceptor can track life cycle of a prom store request and customize or abort its execution at given point.

Required Associated Types

type Error: ErrorExt

Provided Methods

fn pre_write( &self, _write_req: &RowInsertRequests, _ctx: QueryContextRef, ) -> Result<(), Self::Error>

fn pre_read( &self, _read_req: &ReadRequest, _ctx: QueryContextRef, ) -> Result<(), Self::Error>