servers::mysql

Module server

Structs

MysqlServer

MysqlSpawnConfig

MysqlSpawnConfig stores config values which are used to initialize MysqlInstanceShims.

MysqlSpawnRef

MysqlSpawnRef stores arc refs that should be passed to new MysqlInstanceShims.

Constants

DEFAULT_RESULT_SET_WRITE_BUFFER_SIZE 🔒

MYSQL_SERVER