servers
Module query_handler

All query handler traits for various request protocols, like SQL or GRPC.

Instance that wishes to support certain request protocol, just implement the corresponding trait, the Server will handle codec for you.

Note: Query handlers are not confined to only handle read requests, they are expecting to handle write requests too. So the “query” here not might seem ambiguity. However, “query” has been used as some kind of “convention”, it’s the “Q” in “SQL”. So we might better stick to the word “query”.

Modules

grpc

sql

Structs

PromStoreResponse

Traits

InfluxdbLineProtocolHandler

LogQueryHandler

Handle log query requests.

OpenTelemetryProtocolHandler

OpentsdbProtocolHandler

PipelineHandler

PipelineHandler is responsible for handling pipeline related requests.

PromStoreProtocolHandler

Type Aliases

InfluxdbLineProtocolHandlerRef

LogQueryHandlerRef

OpenTelemetryProtocolHandlerRef

OpentsdbProtocolHandlerRef

PipelineHandlerRef

PromStoreProtocolHandlerRef