Module servers::query_handler

All query handler traits for various request protocols, like SQL or GRPC. Instance that wishes to support certain request protocol, just implement the corresponding trait, the Server will handle codec for you.

Note: Query handlers are not confined to only handle read requests, they are expecting to handle write requests too. So the “query” here not might seem ambiguity. However, “query” has been used as some kind of “convention”, it’s the “Q” in “SQL”. So we might better stick to the word “query”.

Modules

grpc

sql

Structs

PromStoreResponse

Traits

InfluxdbLineProtocolHandler

LogHandler

LogHandler is responsible for handling log related requests. It should be able to insert logs and manage pipelines. The pipeline is a series of transformations that can be applied to logs. The pipeline is stored in the database and can be retrieved by name.

OpenTelemetryProtocolHandler

OpentsdbProtocolHandler

PromStoreProtocolHandler

ScriptHandler

Type Aliases

InfluxdbLineProtocolHandlerRef

LogHandlerRef

OpenTelemetryProtocolHandlerRef

OpentsdbProtocolHandlerRef

PromStoreProtocolHandlerRef

ScriptHandlerRef