servers::query_handler
Trait LogHandler

pub trait LogHandler {
    // Required methods
    fn insert_logs<'life0, 'async_trait>(
        &'life0 self,
        log: RowInsertRequests,
        ctx: QueryContextRef,
    ) -> Pin<Box<dyn Future<Output = Result<Output>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_pipeline<'life0, 'life1, 'async_trait>(
        &'life0 self,
        name: &'life1 str,
        version: PipelineVersion,
        query_ctx: QueryContextRef,
    ) -> Pin<Box<dyn Future<Output = Result<Arc<Pipeline<GreptimeTransformer>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn insert_pipeline<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        name: &'life1 str,
        content_type: &'life2 str,
        pipeline: &'life3 str,
        query_ctx: QueryContextRef,
    ) -> Pin<Box<dyn Future<Output = Result<PipelineInfo>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait;
    fn delete_pipeline<'life0, 'life1, 'async_trait>(
        &'life0 self,
        name: &'life1 str,
        version: PipelineVersion,
        query_ctx: QueryContextRef,
    ) -> Pin<Box<dyn Future<Output = Result<Option<()>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

LogHandler is responsible for handling log related requests.

It should be able to insert logs and manage pipelines. The pipeline is a series of transformations that can be applied to logs. The pipeline is stored in the database and can be retrieved by name.

Required Methods

fn insert_logs<'life0, 'async_trait>( &'life0 self, log: RowInsertRequests, ctx: QueryContextRef, ) -> Pin<Box<dyn Future<Output = Result<Output>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn get_pipeline<'life0, 'life1, 'async_trait>( &'life0 self, name: &'life1 str, version: PipelineVersion, query_ctx: QueryContextRef, ) -> Pin<Box<dyn Future<Output = Result<Arc<Pipeline<GreptimeTransformer>>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn insert_pipeline<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 self, name: &'life1 str, content_type: &'life2 str, pipeline: &'life3 str, query_ctx: QueryContextRef, ) -> Pin<Box<dyn Future<Output = Result<PipelineInfo>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

fn delete_pipeline<'life0, 'life1, 'async_trait>( &'life0 self, name: &'life1 str, version: PipelineVersion, query_ctx: QueryContextRef, ) -> Pin<Box<dyn Future<Output = Result<Option<()>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,