servers

Module row_writer

Structs

MultiTableData

TableData

The intermediate data structure for building the write request. It constructs the schema and rows as all input data row parsing is completed.

Enums

TimestampType ğŸ”’

Functions

build_json_column_schema ğŸ”’

check_schema ğŸ”’

check_schema_number ğŸ”’

write_by_schema ğŸ”’

write_by_semantic_type ğŸ”’

write_f64

Write float64 data as a field into the table data.

write_fields

Write data as fields into the table data.

write_json

write_tag

Write data as a tag into the table data.

write_tags

Write data as tags into the table data.

write_ts_to ğŸ”’

write_ts_to_millis

Write timestamp data as milliseconds into the table data.

write_ts_to_nanos

Write timestamp data as nanoseconds into the table data.